04_ai_news_aggr/pyproject.toml
[project]
name = "ai_news_aggr"
version = "0.1.0"
description = "AI news aggregator"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "beautifulsoup4>=4.14.2",
    "docling>=2.61.2",
    "fastapi>=0.110.0",
    "feedparser>=6.0.12",
    "markdown>=3.7.0",
    "markdownify>=0.11.6",
    "openai>=2.7.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.28.0",
    "uvicorn>=0.27.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=8.0.0",
]
